PocketSense/requirements.txt
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.1.0
Pillow==10.4.0 
xmltodict==0.13.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.1.0
django-extensions==3.1.5
ipython==8.12.3


